requirements.txt
Pillow>=10.0.0
easyocr>=1.7.0
opencv-python>=4.8.0
fonttools>=4.43.0
numpy>=1.24.0
scipy>=1.10.0
